Heat III
-인터페이스로 빼둠

+++ src/main/java/com/sturdycobble/createrevision/contents/heatsystem/HeatUtils.java
@@ -2,7 +2,7 @@ package com.sturdycobble.createrevision.contents.heatsystem;
 
 public class HeatUtils {
 	public static void exchangeHeat(HeatContainer from, HeatContainer to) {
-		if ((int)from.getTemp() > (int)to.getTemp()) {
+		if ((int) from.getTemp() > (int) to.getTemp()) {
 			System.out.println("START");
 			System.out.println("    FROM : " + from.getTemp());
 			System.out.println("    TO : " + to.getTemp());
@@ -10,7 +10,7 @@ public class HeatUtils {
 			double transHeat = from.getConductivity() * to.getConductivity() * (from.getTemp() - to.getTemp());
 			double fromTemp = (from.getHeatEnergy() - transHeat) / from.getHeatCapacity();
 			double toTemp = (to.getHeatEnergy() + transHeat) / to.getHeatCapacity();
-			if (fromTemp >= toTemp) {
+			if ((int) fromTemp >= (int) toTemp) {
 				from.setTemp(fromTemp);
 				to.setTemp(toTemp);
 				System.out.println("RESULT : HEAT_TRANSFER");
